+++ RobotProject/src/pl/poznan/put/ioiorobot/motors/Position.java
@@ -2,32 +2,36 @@ package pl.poznan.put.ioiorobot.motors;
 
 import android.graphics.Point;
 
+/**
+ * Klasa przechowująca pozycję robota w układzie współrzędnych wraz z aktualnym
+ * kątem
+ */
 public class Position {
-	private double x;
-	private double y;
-	private double angle;
+	private float x;
+	private float y;
+	private float angle;
 
-	public double x() {
+	public float x() {
 		return x;
 	}
 
-	public double y() {
+	public float y() {
 		return y;
 	}
 
-	public double angle() {
+	public float angle() {
 		return angle;
 	}
-	
+
 	public Point getPoint() {
 		return new Point((int) x, (int) y);
 	}
-	
+
 	public Point getVectorPoint() {
-		return new Point((int) (Math.cos(angle)*100 + x), (int) (Math.sin(angle)*100 + y));
+		return new Point((int) (Math.cos(angle) * 100 + x), (int) (Math.sin(angle) * 100 + y));
 	}
 
-	public void addAngle(double angle) {
+	public void addAngle(float angle) {
 		this.angle += angle;
 		if (this.angle > Math.PI) {
 			this.angle -= 2 * Math.PI;
@@ -37,28 +41,28 @@ public class Position {
 	}
 
 	public Position() {
-		x = y = angle = 0.0;
+		x = y = angle = 0.0f;
 	}
 
-	public Position(double x, double y, double angle) {
+	public Position(float x, float y, float angle) {
 		this.x = x;
 		this.y = y;
 		this.angle = angle;
 	}
-	
+
 	public Position(Position position) {
 		this.x = position.x();
 		this.y = position.y();
 		this.angle = position.angle();
 	}
 
-	public void set(double x, double y, double angle) {
+	public void set(float x, float y, float angle) {
 		this.x = x;
 		this.y = y;
 		this.angle = angle;
 	}
 
-	// public void move(double x, double y, double angle) {
+	// public void move(float x, float y, float angle) {
 	// this.x += x;
 	// this.y += y;
 	// this.angle += angle;
